+++ solar_index/utils.py
